requirements.txt
fastapi
uvicorn
transformers
torch
google-cloud-translate
pdfplumber
pytesseract
Pillow
python-multipart
deep-translator
